Refactor: extract further interfaces, split components
