+++ src/main/reactjs/src/interfaces/selectable.ts
@@ -1,0 +1,4 @@
+export interface Selectable {
+  selected: boolean;
+  toggleSelected(): void;
+}
