+++ src/main/reactjs/src/interfaces/table.ts
@@ -1,0 +1,15 @@
+import { Selectable } from 'interfaces/selectable';
+
+type SetterType<T> = (val: T) => void;
+
+type RowDetailsFn<T> = (details: T, selectionProps: Selectable) => JSX.Element;
+
+export interface PaginatedTableProps<T> {
+  header?: JSX.Element;
+  selectedIndices: Set<number>;
+  setSelectedIndices(selected: SetterType<Set<number>> | Set<number>): void;
+  data: Array<T>;
+  getRowDetails: RowDetailsFn<T>;
+  initialRowsPerPage: number;
+  rowsPerPageOptions: Array<number | { value: number; label: string }>;
+}
